requirements.txt
### ✅ `requirements.txt`

```txt
fastapi
uvicorn
pydantic
python-dotenv
openai>=1.0.0